refactor: change ownership of OutputZoomState
