+++ src/lib.rs
@@ -24,7 +24,6 @@ pub mod screencasting;
 pub mod ui;
 pub mod utils;
 pub mod window;
-pub mod zoom;
 
 #[cfg(test)]
 mod tests;
