+++ src/utils/mod.rs
@@ -45,6 +45,7 @@ pub mod transaction;
 pub mod vblank_throttle;
 pub mod watcher;
 pub mod xwayland;
+pub mod zoom;
 
 pub static IS_SYSTEMD_SERVICE: AtomicBool = AtomicBool::new(false);
