+++ src/utils/zoom.rs
@@ -1,124 +1,6 @@
 use niri_config::ZoomMovementMode;
-use smithay::backend::renderer::element::surface::WaylandSurfaceRenderElement;
-use smithay::backend::renderer::element::utils::{
-    CropRenderElement, Relocate, RelocateRenderElement, RescaleRenderElement,
-};
-use smithay::backend::renderer::element::Element;
 use smithay::output::Output;
 use smithay::utils::{Logical, Physical, Point, Rectangle, Scale, Size};
-
-use crate::layer::mapped::LayerSurfaceRenderElement;
-use crate::layout::tile::TileRenderElement;
-use crate::layout::MonitorRenderElement;
-use crate::niri::PointerRenderElements;
-use crate::niri_render_elements;
-use crate::render_helpers::primary_gpu_texture::PrimaryGpuTextureRenderElement;
-use crate::render_helpers::solid_color::SolidColorRenderElement;
-
-// Define a type alias for the common zoom wrapper: Relocate(Rescale(T))
-pub type ZoomWrapper<T> = RelocateRenderElement<RescaleRenderElement<T>>;
-
-// Unique enum for all zoomed elements - avoids type conflicts with OutputRenderElements since
-// zoomed types are wrapped in a separate enum
-niri_render_elements! {
-    ZoomedRenderElements<R> => {
-        Monitor = ZoomWrapper<MonitorRenderElement<R>>,
-        RescaledTile = ZoomWrapper<RescaleRenderElement<TileRenderElement<R>>>,
-        LayerSurface = ZoomWrapper<LayerSurfaceRenderElement<R>>,
-        RelocatedLayerSurface = ZoomWrapper<CropRenderElement<ZoomWrapper<LayerSurfaceRenderElement<R>>>>,
-        RelocatedColor = ZoomWrapper<CropRenderElement<ZoomWrapper<SolidColorRenderElement>>>,
-        Pointer = ZoomWrapper<PointerRenderElements<R>>,
-        Wayland = ZoomWrapper<WaylandSurfaceRenderElement<R>>,
-        SolidColor = ZoomWrapper<SolidColorRenderElement>,
-        Texture = ZoomWrapper<PrimaryGpuTextureRenderElement>,
-        // We don't apply zoom to WindowMruUi and ExitConfirmDialog elements, so they are intentionally
-        // excluded from this enum to avoid confusion and potential misuse.
-        // ScreenshotUi are handled separately in screenshot_ui module due to their unique
-        // constraints and requirements, so they are also intentionally excluded here.
-    }
-}
-
-/// Wrap an element with the standard zoom transform: Rescale around the focal point, then Relocate
-/// by the subpixel correction. This is the canonical way to apply, except for pointer and
-/// screenshot_ui which are subject to additional constraints and require custom handling.
-pub fn zoom_wrap<E: Element>(
-    elem: E,
-    zoom_factor: f64,
-    output_scale: Scale<f64>,
-    zoom_focal: Point<f64, Logical>,
-) -> ZoomWrapper<E> {
-    let focal_physical: Point<i32, Physical> = zoom_focal.to_physical_precise_round(output_scale);
-    let correction = zoom_subpixel_correction(zoom_focal, zoom_factor, output_scale);
-    RelocateRenderElement::from_element(
-        RescaleRenderElement::from_element(elem, focal_physical, zoom_factor),
-        correction,
-        Relocate::Relative,
-    )
-}
-
-/// Per-output zoom state.
-///
-/// This struct holds the effective zoom values that external consumers (backends,
-/// input, niri rendering) read each frame. Layout writes these values every
-/// animation tick, so they always reflect the current animation/gesture state.
-///
-/// Animation and gesture tracking live in `Monitor` inside the layout module.
-///
-/// Mutable ownership boundaries:
-/// - Layout owns animated `level` / `focal` / `transitioning`.
-/// - Input owns `locked` toggling.
-#[derive(Debug, Clone)]
-pub struct OutputZoomState {
-    /// Current effective zoom level (layout-owned, updated each frame).
-    pub level: f64,
-    /// Current effective focal point in output-local logical coordinates
-    /// (layout-owned, updated each frame).
-    pub focal: Point<f64, Logical>,
-    /// Whether focal point is locked (input-owned toggle).
-    pub locked: bool,
-    /// Whether a zoom animation or gesture is currently in progress (layout-owned).
-    pub transitioning: bool,
-}
-
-impl OutputZoomState {
-    /// Create a new zoom state centered on the given output.
-    pub fn new_for_output(output: &Output) -> Self {
-        let mode_size = output.current_mode().map_or((0, 0).into(), |m| m.size);
-        let scale = output.current_scale().fractional_scale();
-        let logical_size = mode_size.to_f64().to_logical(scale);
-        Self {
-            level: 1.0,
-            focal: Point::from((logical_size.w / 2.0, logical_size.h / 2.0)),
-            locked: false,
-            transitioning: false,
-        }
-    }
-
-    pub fn is_active(&self) -> bool {
-        self.level > 1.0
-    }
-
-    pub fn viewport_global(
-        &self,
-        output_geometry: Rectangle<f64, Logical>,
-    ) -> Rectangle<f64, Logical> {
-        let focal_global = self.focal + output_geometry.loc;
-        apply_zoom_viewport(output_geometry, focal_global, self.level)
-    }
-
-    pub fn clamp_to_viewport(
-        &self,
-        pos: Point<f64, Logical>,
-        output_geometry: Rectangle<f64, Logical>,
-    ) -> Point<f64, Logical> {
-        let vp = self.viewport_global(output_geometry);
-        Point::from((
-            pos.x.clamp(vp.loc.x, vp.loc.x + vp.size.w - f64::EPSILON),
-            pos.y.clamp(vp.loc.y, vp.loc.y + vp.size.h - f64::EPSILON),
-        ))
-    }
-}
-
 pub fn apply_zoom_viewport(
     mut output_rect: Rectangle<f64, Logical>,
     focal_point: Point<f64, Logical>,
